+++ README.md
@@ -63,8 +63,6 @@ results = gilda.annotate('Calcium is released from the ER.')
 
 ### Use as a web service
 
-To use Gilda's REST service, you will need to install with the `[web]` extra
-as in `pip install gilda[web]` or locally `pip install -e .[web]`.
 The REST service accepts POST requests with a JSON header on the `/ground`
 endpoint. There is a public REST service running at http://grounding.indra.bio
 but the service can also be run locally as
