+++ setup.py
@@ -36,18 +36,17 @@ setup(name='gilda',
           'Programming Language :: Python :: 3.11',
       ],
       packages=find_packages(),
-      install_requires=['regex', 'adeft>=0.11.0', 'boto3',
-                        'pystow>=0.1.10', 'unidecode'],
+      install_requires=['regex', 'adeft>=0.11.0', 'boto3', 'flask',
+                        'flask-restx', 'pystow>=0.1.10', 'unidecode',
+                        'werkzeug<2.2',
+                        ],
       extras_require={'test': ['pytest', 'pytest-cov', 'pandas'],
                       'terms': ['indra', 'obonet'],
                       'benchmarks': ['pandas>=1.0', 'requests',
                                      'tabulate', 'tqdm', 'click'],
-                      'web': [
-                        'werkzeug<2.2',
-                        'flask',
+                      'ui': [
                         'flask-wtf',
                         'flask-bootstrap',
-                        'flask-restx',
                       ],
                       'docs': [
                           "sphinx",
